added documentation and moved configuration filclass into provider class
