--- sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/providers/BigQueryDirectReadSchemaTransformConfiguration.java
@@ -1,69 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.beam.sdk.io.gcp.bigquery.providers;
-
-import com.google.auto.value.AutoValue;
-import java.util.List;
-import javax.annotation.Nullable;
-import org.apache.beam.sdk.annotations.Experimental;
-import org.apache.beam.sdk.io.gcp.bigquery.BigQueryServices;
-import org.apache.beam.sdk.schemas.AutoValueSchema;
-import org.apache.beam.sdk.schemas.annotations.DefaultSchema;
-
-@Experimental
-@DefaultSchema(AutoValueSchema.class)
-@AutoValue
-public abstract class BigQueryDirectReadSchemaTransformConfiguration {
-
-  /** Instantiates a {@link BigQueryDirectReadSchemaTransformConfiguration.Builder} instance. */
-  public static Builder builder() {
-    return new AutoValue_BigQueryDirectReadSchemaTransformConfiguration.Builder();
-  }
-
-  @Nullable
-  public abstract String getQuery();
-
-  @Nullable
-  public abstract String getTableSpec();
-
-  @Nullable
-  public abstract String getRowRestriction();
-
-  @Nullable
-  public abstract List<String> getSelectedFields();
-
-  @Nullable
-  public abstract BigQueryServices getBigQueryServices();
-
-  /** Builder for the {@link BigQueryDirectReadSchemaTransformConfiguration}. */
-  @AutoValue.Builder
-  public abstract static class Builder {
-    public abstract Builder setQuery(String query);
-
-    public abstract Builder setTableSpec(String tableSpec);
-
-    public abstract Builder setRowRestriction(String rowRestriction);
-
-    public abstract Builder setSelectedFields(List<String> selectedFields);
-
-    public abstract Builder setBigQueryServices(BigQueryServices bigQueryServices);
-
-    /** Builds a {@link BigQueryDirectReadSchemaTransformConfiguration} instance. */
-    public abstract BigQueryDirectReadSchemaTransformConfiguration build();
-  }
-}
